add support for printing the stencil
I did this so we can eventually print filament that is conductive on top of
the 3D printed board.  The theory I have is that I will print a 3D board, add
all the components to it, and then 3D print the stencil with conductive PLA
which will act as the solder.